Correct return values in getRequiredSkillTries
This correction was at the request of PR # 3217, the code was inspired directly from the repository of https://github.com/SaiyansKing/optimized_forgottenserver/commit/a5f525ac1799f0ff2ce4afa1aedba34fc2faaaa0

+++ src/vocation.h
@@ -100,8 +100,8 @@ class Vocation
 		std::string name = "none";
 		std::string description;
 
-		float skillMultipliers[SKILL_LAST + 1] = {1.5f, 2.0f, 2.0f, 2.0f, 2.0f, 1.5f, 1.1f};
-		float manaMultiplier = 4.0f;
+		double skillMultipliers[SKILL_LAST + 1] = {1.5, 2.0, 2.0, 2.0, 2.0, 1.5, 1.1};
+		double manaMultiplier = 4.0;
 
 		uint32_t gainHealthTicks = 6;
 		uint32_t gainHealthAmount = 1;
